+++ js/data/game.js
@@ -5,35 +5,6 @@ export const INITIAL_STATE = Object.freeze({
   answers: [],
 });
 
-const images = {
-  paintings: [
-    // People
-    `https://k42.kn3.net/CF42609C8.jpg`,
-
-    // Animals
-    `https://k42.kn3.net/D2F0370D6.jpg`,
-
-    // Nature
-    `https://k32.kn3.net/5C7060EC5.jpg`
-  ],
-  photos: [
-    // People
-    `http://i.imgur.com/1KegWPz.jpg`,
-
-    // Animals
-    `https://i.imgur.com/DiHM5Zb.jpg`,
-
-    // Nature
-    `http://i.imgur.com/DKR1HtB.jpg`
-  ]
-};
-
-const newImages = {
-  paintings: images.paintings.map(it=>{
-
-  })
-}
-
 export const PIXEL_HUNTER = [
   {
     photoPaint: true,
