+++ js/game-render.js
@@ -3,7 +3,6 @@ import {renderGameTwo} from "./game-2";
 import {renderGameThree} from "./game-3";
 import {renderGameOne} from "./game-1";
 import {showStats} from "./stats";
-import {resize} from "./data/resize";
 
 export const setQuestionStyle = (state) => {
   const answersLength = PIXEL_HUNTER[state.level].answers.length;
@@ -36,18 +35,11 @@ export const shouldLevelRender = (state, answer = -1, lives = state.lives) => {
   }
 };
 
-export const gameRender = (level, width, height) => {
+export const gameRender = (level) => {
   return `
-      ${level.answers.map((it, i) => {
-    const img = new Image();
-    img.addEventListener(`load`, function () {
-      const naturalSize = {width: img.width, height: img.height};
-      const resolution = resize({width, height}, naturalSize);
-      img.width = resolution.width;
-      img.height = resolution.height;
-      return level.photoPaint
-        ? `<div class="game__option">
-              <img src=${img.src} alt="Option ${i + 1}" width=${img.width} height=${img.height}>
+    ${level.answers.map((it, i) => {
+    return level.photoPaint
+      ? `<div class="game__option">      
               <label class="game__answer game__answer--photo">
                 <input class="visually-hidden" name="question${i + 1}" type="radio" value="photo">
                 <span>Фото</span>
@@ -57,10 +49,7 @@ export const gameRender = (level, width, height) => {
                 <span>Рисунок</span>
               </label>
             </div>`
-        : `<div class="game__option">
-              <img src=${img.src} alt="Option ${i + 1}" width=${img.width} height=${img.height}>
-              </div>`;
-    });
-    img.src = it.img;
+      : `<div class="game__option">            
+           </div>`;
   }).join(``)}`;
 };
